Move view model creation into TokenList
